api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
aiofiles==24.1.0
redis==5.2.1
Pillow==11.1.0
rembg==2.0.62
onnxruntime==1.20.1
pydantic==2.10.5
pydantic-settings==2.7.1
sse-starlette==2.2.1
opencv-python-headless==4.9.0.80
numpy==1.26.4
